+++ packages/frameworks/vue/src/date-picker/date-picker.tsx
@@ -22,7 +22,7 @@ export const DatePicker = defineComponent<DatePickerProps>(
       lazyMount: props.lazyMount,
       unmountOnExit: props.unmountOnExit,
     }))
-    const presenceApi = usePresence(presenceProps as any, emit)
+    const presenceApi = usePresence(presenceProps, emit)
 
     DatePickerProvider(api)
     PresenceProvider(presenceApi)