+++ packages/frameworks/vue/src/select/use-select.ts
@@ -59,8 +59,7 @@ export const useSelect = <T extends CollectionItem>(
   watch(
     () => props.items,
     (value) => {
-      // FIXME: blocked by zagjs setItems types
-      api.value.collection.setItems(value as any)
+      api.value.collection.setItems(value)
     },
   )