+++ packages/scripts/src/generate-stories.ts
@@ -25,7 +25,7 @@ const extractScriptSetup = (fileContent: string): string | null => {
   if (match) {
     return match[1]
       .trim()
-      .replace(/import\s+\{([^\}]+)\}\s+from\s+'\.\/'/g, "import {$1} from '@ark-ui/vue'")
+      .replace(/import\s+\{([^\}]+)\}\s+from\s+'\.\.\/?'/g, "import {$1} from '@ark-ui/vue'")
       .replace(
         /import\s+\{\sPortal\s\}\sfrom\s'..\/portal'/g,
         "import { Portal } from '@ark-ui/vue'",
@@ -79,11 +79,11 @@ const main = async () => {
         const importText = importDecl
           .getText()
           .replace(
-            /import\s+\{([^\}]+)\}\s+from\s+'\.\/'/g,
+            /import\s+\{([^\}]+)\}\s+from\s+'\.\.\/?'/g,
             `import {$1} from '@ark-ui/${framework}'`,
           )
           .replace(
-            /import\s+\{\sPortal\s\}\sfrom\s'..\/portal'/g,
+            /import\s+\{\sPortal\s\}\sfrom\s'..\/..\/portal'/g,
             `import { Portal } from '@ark-ui/${framework}'`,
           )
           .replace(/@zag-js\/react/g, '@ark-ui/react')